Moving ringbuffer one folder level up.
